Refactor NPE-C loss logic to Strategy pattern
